refactored the view model
