--- app/src/main/java/me/simonpojok/sportlive/ui/common/ViewState.kt
@@ -1,3 +1,0 @@
-package me.simonpojok.sportlive.ui.common
-
-interface ViewState
