+++ app/src/main/java/me/simonpojok/sportlive/ui/common/viewmodel/PresentationDestination.kt
@@ -1,0 +1,3 @@
+package me.simonpojok.sportlive.ui.common.viewmodel
+
+interface PresentationDestination
