+++ app/src/main/java/me/simonpojok/sportlive/ui/common/viewmodel/exception/DomainException.kt
@@ -1,4 +1,4 @@
-package me.simonpojok.sportlive.ui.common.exception
+package me.simonpojok.sportlive.ui.common.viewmodel.exception
 
 abstract class DomainException(open val throwable: Throwable) : Exception(throwable) {
     constructor(message: String) : this(Exception(message))
