+++ app/src/main/java/me/simonpojok/sportlive/ui/common/viewmodel/exception/PresentationException.kt
@@ -1,3 +1,3 @@
-package me.simonpojok.sportlive.ui.common.exception
+package me.simonpojok.sportlive.ui.common.viewmodel.exception
 
 abstract class PresentationException(val throwable: Throwable = Throwable()) : Exception(throwable)
